# docs/swagger/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: earthcake2233
    url: https://github.com/earthcake2233
  description: cakecake - 轻量级弹幕视频分享平台后端 API
  license:
    name: Custom Non-Commercial License
    url: https://github.com/earthcake2233/cakecake/blob/main/LICENSE
  termsOfService: https://github.com/earthcake2233/cakecake/blob/main/LICENSE
  title: cakecake API
  version: "1.0"
paths:
...
  /dm/conversations:
    get:
      description: Get all direct message conversations for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List DM conversations
      tags:
      - DM
    post:
      description: Start a new direct message conversation with another user
      parameters:
      - description: Other participant user ID
        in: body
        name: body
        required: true
        schema:
          properties:
            participant_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create a DM conversation
      tags:
      - DM
  /dm/conversations/{id}:
    delete:
      description: Remove a direct message conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete a DM conversation
      tags:
      - DM
  /dm/conversations/{id}/messages:
    get:
      description: Get paginated messages for a DM conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List messages in a conversation
      tags:
      - DM
    post:
      description: Post a new message to a DM conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message content
        in: body
        name: body
        required: true
        schema:
          properties:
            content:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Send a message in a conversation
      tags:
      - DM
  /dm/conversations/{id}/reset:
    post:
      description: Clear the conversation history for an AI agent DM
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Reset AI agent conversation
      tags:
      - DM
  /dm/conversations/{id}/settings:
    patch:
      description: Update settings (e.g. agent config) for a conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Settings
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update DM conversation settings
      tags:
      - DM
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token, e.g. "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"